test(ai): isolate storage coverage in its ownership directory
